% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isorunner.R
\docType{package}
\name{isorunner-package}
\alias{isorunner}
\alias{isorunner-package}
\title{isorunner: Process stable isotope data}
\description{
A package to process stable isotope ratio mass spectrometry data in
a fully transparent and reproducible manner that is easy to use and communicate.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kopflab/isorunner}
  \item Report bugs at \url{https://github.com/kopflab/isorunner/issues}
}

}
\keyword{internal}
